CTP/HuggingFace/Cluster_Experiments/requirements.txt
gradio
sentence-transformers
umap-learn
hdbscan  # Note: can also install via conda: conda install -c conda-forge hdbscan
scikit-learn
numpy
matplotlib
nltk
pyclustering

